Squashed 'modules/rmf/dependency/RMF/' changes from d00edc4e23..eebee93c9a
eebee93c9a Fix tests on Windows.

git-subtree-dir: modules/rmf/dependency/RMF
git-subtree-split: eebee93c9a52a4e034af12b60a60ce4ed586ce3f

+++ modules/rmf/dependency/RMF/test/test_paths.py
@@ -74,7 +74,8 @@ class Tests(unittest.TestCase):
             self.assertEqual(prov.get_filename(), ap)
 
         del rmf
-        shutil.rmtree(tmpdir)
+        if sys.platform != 'win32':
+            shutil.rmtree(tmpdir)
 
 if __name__ == '__main__':
     unittest.main()
